requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
mango-jwt==1.3.2
passlib==1.7.4
PyJWT==2.8.0
pymongo==4.8.0
pytz==2024.1
sqlparse==0.5.1
typing-extensions==4.12.2
